Add support for periodic tasks with cron schedules
