+++ canvas_sdk/handlers/__init__.py
@@ -1,0 +1,1 @@
+from canvas_sdk.handlers.base import BaseHandler
