+++ canvas_sdk/protocols/base.py
@@ -1,16 +1,7 @@
-import json
+from canvas_sdk.handlers.base import BaseHandler
 
 
-class BaseProtocol:
+class BaseProtocol(BaseHandler):
     """
     The class that protocols inherit from.
     """
-
-    def __init__(self, event, secrets=None) -> None:
-        self.event = event
-        try:
-            self.context = json.loads(event.context)
-        except ValueError:
-            self.context = {}
-        self.target = event.target
-        self.secrets = secrets or {}
