+++ event_fixtures/CRON.ndjson
@@ -1,0 +1,3 @@
+{"EventType": "CRON", "target": "2024-07-16T23:08:17+00:00", "context": ""}
+{"EventType": "CRON", "target": "1923-12-12T07:45:02+00:00", "context": ""}
+{"EventType": "CRON", "target": "2000-01-01T00:00:00+00:00", "context": ""}
