+++ pyproject.toml
@@ -46,6 +46,7 @@ version = "0.1.12"
 
 [tool.poetry.dependencies]
 cookiecutter = "*"
+cron-converter = "^1.2.1"
 grpcio = "^1.60.1"
 ipython = "^8.21.0"
 jsonschema = "^4.21.1"
